refactor: editor plugins📦
